[008] Add (dysfunctional) Terraform config
This commit adds a dysfunctional set of Terraform config files which
were used in an attempt to automate creation of storage integration
and the IAM role that provides access to Snowflake.

The hope was that Terraform provider for Snowflake has already resoved
the cyclic dependency between the storage integration and the IAM role
that facilitates the S3 access to Snowflake's AWS user, but that issue
isn't resolved yet. Therefore the config fails with an "Error: Cycle"
in the plan phase.