[009] Use Terraform to setup stage
This commit changes the formerly dysfunctional terraform code so that
it can setup Snowflake Stage object using an AWS IAM user and its
`access_key\secret_key`.

The Terraform config creates an IAM user using a permission policy
which only allows it to read objects from the S3 bucket which hosts
the data files, and generates the Snowflake Stage using the same
credentials.

During the process it creates the database and schemas.